Fix compiler errors/warnings
- Activate several common build flag (check CMake file in src)
- Fix all compiler errors/warnings

+++ inc/utilities/hexdump/ez_hexdump.h
@@ -80,7 +80,7 @@ extern "C" {
 * @endcode
 *
 *****************************************************************************/
-void ezHexdump(void *address, uint16_t size);
+void ezHexdump(void *address, uint32_t size);
 
 #endif /* CONFIG_HELPER_HEXDUMP */
